atualização no findAll para trazer mais informações

+++ src/main/java/com/dchristofolli/projects/awss3/model/FileModel.java
@@ -9,5 +9,5 @@ import lombok.Data;
 @Data
 public class FileModel {
     private String fileName;
-    private long size;
+    private Integer size;
 }
